+++ Sources/NablaCore/Data/Authentication/Authenticator.swift
@@ -7,6 +7,8 @@ public enum AuthenticationState {
 
 // sourcery: AutoMockable
 public protocol Authenticator {
+    var currentUserId: String? { get }
+    
     func authenticate(userId: String)
     func logOut()
     func markTokensAsInvalid()